src/quote_generator/requirements-quote-generator.txt
# Optional dependency. Pillow enables PNG output. The script works without it and will always write an SVG.
Pillow>=9.0.0
